fix(mdxEditor): refactor markdown helpers
